+++ CHANGELOG.md
@@ -386,7 +386,7 @@
   * tpm2_create: -g/-G become optional options.
   * tpm2_createprimary: -g/-G become optional options.
   * tpm2_verifysignature - Option `-r` changes to `-f` and supports signature format "rsa".
-  * tpm2_import - Parent public data option, `-K` is optional.
+  * tpm2_import - Parent public data option is optional and changes from `-K` to `-U`.
   * tpm2_import - Supports importing external RSA 2048 keys via pem files.
   * tpm2_pcrlist: Option `--algorithm` changes to `--halg`, which is in line with other tools.
   * tpm2_verifysignature: Option `-r` and `--raw` have been removed. This were unused within the tool.